+++ CHANGELOG.md
@@ -1,3 +1,25 @@
+## [144.17.0](https://github.com/Sage/carbon/compare/v144.16.0...v144.17.0) (2024-12-10)
+
+
+### Features
+
+* **menu-fullscreen:** fix colours when menuitem used as child of menusegmenttitle ([7b6ae46](https://github.com/Sage/carbon/commit/7b6ae46056ebc919eb60d8895f5d6cb461379dc8))
+
+## [144.16.0](https://github.com/Sage/carbon/compare/v144.15.0...v144.16.0) (2024-12-09)
+
+
+### Features
+
+* **classname prop:** add deprecation warning to indicate impending removal of className prop ([d838a08](https://github.com/Sage/carbon/commit/d838a082859c5431490127376bcda8baec73576f))
+* **tests:** add deprecation message test coverage ([0fc4090](https://github.com/Sage/carbon/commit/0fc4090298e2dbdf93538b8fc9ce95eb23e9ed21))
+
+## [144.15.0](https://github.com/Sage/carbon/compare/v144.14.0...v144.15.0) (2024-12-06)
+
+
+### Features
+
+* **decimal:** add a deprecation warning for onKeyPress prop ([8e297e6](https://github.com/Sage/carbon/commit/8e297e6311b908b33a952108386be2976d4bd11e))
+
 ## [144.14.0](https://github.com/Sage/carbon/compare/v144.13.0...v144.14.0) (2024-12-06)